requirements.txt
kafka-python==1.4.6
newspaper3k==0.2.8
pip-chill==0.1.8
pymongo==3.9.0
redis==3.3.8
